Move GM to server side
This GM is a server handled GM, not client sided as seen [here](https://github.com/DarkflameUniverse/DarkflameServer/blob/09dfb6df3a53ce3fd3870351e8b3e6633e5a349b/dGame/dGameMessages/GameMessageHandler.cpp#L652-L654)
